+++ lib/src/navigation_guard.dart
@@ -12,7 +12,7 @@ class VNavigationGuard extends StatefulWidget {
   /// associated to this [VNavigationGuard] is in the previous route
   /// but not in the new one
   ///
-  /// Use [vRedirector] if you want to redirect.
+  /// Use [vRedirector] if you want to redirect or stop the navigation.
   /// DO NOT use VRouterData methods to redirect.
   /// [vRedirector] also has information about the route you leave and the route you go to
   ///
@@ -64,11 +64,14 @@ class VNavigationGuard extends StatefulWidget {
   final void Function(BuildContext context, String from, String to) afterUpdate;
 
   /// Called when a pop event occurs.
-  /// A pop event can be called programmatically (with
-  /// [VRouterData.of(context).pop()]) or by other widgets
-  /// such as the appBar back button
+  /// A pop event can be called programmatically (with [VRouterData.of(context).pop()])
+  /// or by other widgets such as the appBar back button
   ///
-  /// return false to stop the pop event
+  /// Use [vRedirector] if you want to redirect or stop the navigation.
+  /// DO NOT use VRouterData methods to redirect.
+  /// [vRedirector] also has information about the route you leave and the route you go to
+  ///
+  /// The route you go to is calculated based on [VRouterState._defaultPop]
   ///
   /// Note that you should consider the pop cycle to
   /// handle this precisely, see [https://vrouter.dev/guide/Advanced/Pop%20Events/onPop]
@@ -77,12 +80,16 @@ class VNavigationGuard extends StatefulWidget {
   ///   * [VRouter.onPop] for global level onPop
   ///   * [VRouteElement.onPop] for route level onPop
   ///   * [VRouterState._defaultPop] for the default onPop
-  final Future<bool> Function(BuildContext context) onPop;
+  final Future<void> Function(VRedirector vRedirector) onPop;
 
-  /// Called when a system pop event occurs. This happens on android
-  /// when the system back button is pressed
+  /// Called when a system pop event occurs.
+  /// This happens on android when the system back button is pressed
   ///
-  /// return false to stop the pop event
+  /// Use [vRedirector] if you want to redirect or stop the navigation.
+  /// DO NOT use VRouterData methods to redirect.
+  /// [vRedirector] also has information about the route you leave and the route you go to
+  ///
+  /// The route you go to is calculated based on [VRouterState._defaultPop]
   ///
   /// Note that you should consider the systemPop cycle to
   /// handle this precisely, see [https://vrouter.dev/guide/Advanced/Pop%20Events/onSystemPop]
@@ -90,7 +97,7 @@ class VNavigationGuard extends StatefulWidget {
   /// Also see:
   ///   * [VRouter.onSystemPop] for global level onSystemPop
   ///   * [VRouteElement.onSystemPop] for route level onSystemPop
-  final Future<bool> Function(BuildContext context) onSystemPop;
+  final Future<void> Function(VRedirector vRedirector) onSystemPop;
 
   const VNavigationGuard({
     Key key,
@@ -109,12 +116,13 @@ class VNavigationGuard extends StatefulWidget {
 class _VNavigationGuardState extends State<VNavigationGuard> {
   @override
   void initState() {
-    VNavigationGuardMessage(vNavigationGuard: widget, localContext: context).dispatch(context);
+    VNavigationGuardMessage(vNavigationGuard: widget, localContext: context)
+        .dispatch(context);
     super.initState();
     if (widget.afterEnter != null) {
       WidgetsBinding.instance.addPostFrameCallback((_) {
-        widget.afterEnter(
-            context, VRouterData.of(context).previousUrl, VRouterData.of(context).url);
+        widget.afterEnter(context, VRouterData.of(context).previousUrl,
+            VRouterData.of(context).url);
       });
     }
   }
@@ -124,7 +132,8 @@ class _VNavigationGuardState extends State<VNavigationGuard> {
   // are necessary when changes to VNavigationGuard are made
   @override
   void reassemble() {
-    VNavigationGuardMessage(vNavigationGuard: widget, localContext: context).dispatch(context);
+    VNavigationGuardMessage(vNavigationGuard: widget, localContext: context)
+        .dispatch(context);
     super.reassemble();
   }
 
@@ -140,5 +149,6 @@ class VNavigationGuardMessage extends Notification {
   final VNavigationGuard vNavigationGuard;
   final BuildContext localContext;
 
-  VNavigationGuardMessage({@required this.vNavigationGuard, @required this.localContext});
+  VNavigationGuardMessage(
+      {@required this.vNavigationGuard, @required this.localContext});
 }