+++ lib/src/router.dart
@@ -21,9 +21,8 @@ class VRouter extends StatefulWidget {
   /// Note that if this is not implemented, every route which does not implement
   /// its own buildTransition will be given a default transition: this of a
   /// [MaterialPage]
-  final Widget Function(
-          Animation<double> animation, Animation<double> secondaryAnimation, Widget child)
-      buildTransition;
+  final Widget Function(Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) buildTransition;
 
   /// The duration of [VRouter.buildTransition]
   final Duration transitionDuration;
@@ -39,7 +38,7 @@ class VRouter extends StatefulWidget {
   final VRouterModes mode;
 
   /// Called when a url changes, before the url is updated
-  /// Use [vRedirector] if you want to redirect.
+  /// Use [vRedirector] if you want to redirect or stop the navigation.
   /// DO NOT use VRouterData methods to redirect.
   /// [vRedirector] also has information about the route you leave and the route you go to
   ///
@@ -62,7 +61,7 @@ class VRouter extends StatefulWidget {
 
   /// This is called before the url is updated but after all beforeLeave are called
   ///
-  /// Use [vRedirector] if you want to redirect.
+  /// Use [vRedirector] if you want to redirect or stop the navigation.
   /// DO NOT use VRouterData methods to redirect.
   /// [vRedirector] also has information about the route you leave and the route you go to
   ///
@@ -87,11 +86,14 @@ class VRouter extends StatefulWidget {
   final void Function(BuildContext context, String from, String to) afterEnter;
 
   /// Called after the [VRouteElement.onPopPage] when a pop event occurs
-  /// You can use the context to call [VRouterData.of(context).push]
-  /// or [VRouterData.of(context).pushNamed], if you do return true.
-  /// Return true if you handled the event, false otherwise
-  /// Note that returning false will trigger the default behaviour
-  /// If [onSystemPop] is null, system event will trigger this method as well
+  /// A pop event can be called programmatically (with [VRouterData.of(context).pop()])
+  /// or by other widgets such as the appBar back button
+  ///
+  /// Use [vRedirector] if you want to redirect or stop the navigation.
+  /// DO NOT use VRouterData methods to redirect.
+  /// [vRedirector] also has information about the route you leave and the route you go to
+  ///
+  /// The route you go to is calculated based on [VRouterState._defaultPop]
   ///
   /// Note that you should consider the pop cycle to
   /// handle this precisely, see [https://vrouter.dev/guide/Advanced/Pop%20Events/onPop]
@@ -100,14 +102,16 @@ class VRouter extends StatefulWidget {
   ///   * [VRouteElement.onPop] for route level onPop
   ///   * [VNavigationGuard.onPop] for widget level onPop
   ///   * [VRouterState._defaultPop] for the default onPop
-  final Future<bool> Function(BuildContext context) onPop;
+  final Future<void> Function(VRedirector vRedirector) onPop;
 
   /// Called after the [VRouteElement.onPopPage] when a system pop event occurs.
   /// This happens on android when the system back button is pressed.
-  /// You can use the context to call [VRouterData.of(context).push]
-  /// or [VRouterData.of(context).pushNamed], if you do return true.
   ///
-  /// Return true if you handled the event, false otherwise
+  /// Use [vRedirector] if you want to redirect or stop the navigation.
+  /// DO NOT use VRouterData methods to redirect.
+  /// [vRedirector] also has information about the route you leave and the route you go to
+  ///
+  /// The route you go to is calculated based on [VRouterState._defaultPop]
   ///
   /// Note that you should consider the systemPop cycle to
   /// handle this precisely, see [https://vrouter.dev/guide/Advanced/Pop%20Events/onSystemPop]
@@ -115,7 +119,7 @@ class VRouter extends StatefulWidget {
   /// Also see:
   ///   * [VRouteElement.onSystemPop] for route level onSystemPop
   ///   * [VNavigationGuard.onSystemPop] for widget level onSystemPop
-  final Future<bool> Function(BuildContext context) onSystemPop;
+  final Future<void> Function(VRedirector vRedirector) onSystemPop;
 
   VRouter({
     Key key,
@@ -577,12 +581,14 @@ class VRouterState extends State<VRouter> {
   @override
   Widget build(BuildContext context) {
     return SimpleUrlHandler(
-      urlToAppState: (BuildContext context, RouteInformation routeInformation) async {
+      urlToAppState:
+          (BuildContext context, RouteInformation routeInformation) async {
         if (routeInformation.location != null && !ignoreNextBrowserCalls) {
           // Get the new state
           final newState = (kIsWeb)
-              ? Map<String, String>.from(jsonDecode(routeInformation.state as String ??
-                  (BrowserHelpers.getHistoryState() ?? '{}')))
+              ? Map<String, String>.from(jsonDecode(
+                  routeInformation.state as String ??
+                      (BrowserHelpers.getHistoryState() ?? '{}')))
               : <String, String>{};
 
           // Get the new serial count
@@ -610,8 +616,9 @@ class VRouterState extends State<VRouter> {
             '-2': _historyState,
             '-1': vRoute?.key?.currentState?.historyState,
             for (var pages in flattenPages)
-              '${pages.child.depth}': pages.child.stateKey?.currentState?.historyState ??
-                  pages.child.initialHistorySate,
+              '${pages.child.depth}':
+                  pages.child.stateKey?.currentState?.historyState ??
+                      pages.child.initialHistorySate,
           }),
         );
       },
@@ -682,13 +689,14 @@ class VRouterState extends State<VRouter> {
   }) {
     final routesFlatten = <_VRoutePath>[];
     final parentPath = parentVRoutePath?.path ?? '';
-    var parentVRouteElements =
-        List<VRouteElement>.from(parentVRoutePath?.vRouteElements ?? <VRouteElement>[]);
+    var parentVRouteElements = List<VRouteElement>.from(
+        parentVRoutePath?.vRouteElements ?? <VRouteElement>[]);
 
     // For each childRoutes
     for (var childRoute in childRoutes) {
       // Add the VRouteElement to the parent ones to from the VRouteElements list
-      final vRouteElements = List<VRouteElement>.from([...parentVRouteElements, childRoute]);
+      final vRouteElements =
+          List<VRouteElement>.from([...parentVRouteElements, childRoute]);
 
       // If the path is null, just get the route from the subroutes
       if (childRoute.path == null) {
@@ -718,7 +726,8 @@ class VRouterState extends State<VRouter> {
 
         // Get the pathRegExp and the new parameters
         var newGlobalParameters = <String>[];
-        final globalPathRegExp = pathToRegExp(globalPath, parameters: newGlobalParameters);
+        final globalPathRegExp =
+            pathToRegExp(globalPath, parameters: newGlobalParameters);
 
         // Instantiate the new vRoutePath
         final vRoutePath = _VRoutePath(
@@ -738,11 +747,13 @@ class VRouterState extends State<VRouter> {
 
           for (var alias in childRoute.aliases) {
             // Get the global path
-            final globalPath = (alias.startsWith('/')) ? alias : parentPath + '/$alias';
+            final globalPath =
+                (alias.startsWith('/')) ? alias : parentPath + '/$alias';
 
             // Get the pathRegExp and the new parameters
             var newGlobalParameters = <String>[];
-            final globalPathRegExp = pathToRegExp(globalPath, parameters: newGlobalParameters);
+            final globalPathRegExp =
+                pathToRegExp(globalPath, parameters: newGlobalParameters);
 
             // Instantiate the new vRoutePath
             final vRoutePath = _VRoutePath(
@@ -834,15 +845,18 @@ class VRouterState extends State<VRouter> {
   }) {
     // Find the path corresponding to the name
     var newPath = pathToRoutes
-        .firstWhere((_VRoutePath vRoutePathRegexp) => (vRoutePathRegexp.name == name),
-            orElse: () => throw Exception('Could not find [VRouteElement] with name $name'))
+        .firstWhere(
+            (_VRoutePath vRoutePathRegexp) => (vRoutePathRegexp.name == name),
+            orElse: () => throw Exception(
+                'Could not find [VRouteElement] with name $name'))
         .path;
 
     // Inject the given path parameters into the new path
     newPath = pathToFunction(newPath)(pathParameters);
 
     // Update the url with the found and completed path
-    _updateUrl(newPath, queryParameters: queryParameters, isReplacement: isReplacement);
+    _updateUrl(newPath,
+        queryParameters: queryParameters, isReplacement: isReplacement);
   }
 
   /// Recursive function which builds a nested representation of the given route
@@ -880,14 +894,17 @@ class VRouterState extends State<VRouter> {
       var localPath = vRouteElement.path;
 
       // First remove any / that would be in first position
-      if (remainingUrl.startsWith('/')) remainingUrl = remainingUrl.replaceFirst('/', '');
-      if (localPath.startsWith('/')) localPath = localPath.replaceFirst('/', '');
+      if (remainingUrl.startsWith('/'))
+        remainingUrl = remainingUrl.replaceFirst('/', '');
+      if (localPath.startsWith('/'))
+        localPath = localPath.replaceFirst('/', '');
 
       // We try to match the pathRegExp with the remainingUrl
       // This is null if a deeper-nester VRouteElement has a path
       // which starts with '/' (which means that this pathRegExp is not part
       // of the url)
-      final match = vRouteElement.pathRegExp.matchAsPrefix(Uri.decodeComponent(remainingUrl));
+      final match = vRouteElement.pathRegExp
+          .matchAsPrefix(Uri.decodeComponent(remainingUrl));
 
       // If the previous match didn't fail, we get the remainingUrl be stripping of the
       // part of the url which matched
@@ -910,13 +927,17 @@ class VRouterState extends State<VRouter> {
       final newVPage = VPage(
         key: vRouteElement.key ?? ValueKey(vRouteElement.path),
         name: vRouteElement.name ?? vRouteElement.path,
-        buildTransition: vRouteElement.buildTransition ?? widget.buildTransition,
-        transitionDuration: vRouteElement.transitionDuration ?? widget.transitionDuration,
-        reverseTransitionDuration:
-            vRouteElement.reverseTransitionDuration ?? widget.reverseTransitionDuration,
+        buildTransition:
+            vRouteElement.buildTransition ?? widget.buildTransition,
+        transitionDuration:
+            vRouteElement.transitionDuration ?? widget.transitionDuration,
+        reverseTransitionDuration: vRouteElement.reverseTransitionDuration ??
+            widget.reverseTransitionDuration,
         child: RouteElementWidget(
           stateKey: vRouteElement.stateKey,
-          child: vRouteElement.widget ?? Builder(builder: (context) => vRouteElement.widgetBuilder(context)),
+          child: vRouteElement.widget ??
+              Builder(
+                  builder: (context) => vRouteElement.widgetBuilder(context)),
           depth: index,
           pathParameters: localParameters,
           name: vRouteElement.name,
@@ -973,13 +994,17 @@ class VRouterState extends State<VRouter> {
       final newVPage = VPage(
         key: vRouteElement.key ?? ValueKey(vRouteElement.path),
         name: vRouteElement.name ?? vRouteElement.path,
-        buildTransition: vRouteElement.buildTransition ?? widget.buildTransition,
-        transitionDuration: vRouteElement.transitionDuration ?? widget.transitionDuration,
-        reverseTransitionDuration:
-            vRouteElement.reverseTransitionDuration ?? widget.reverseTransitionDuration,
+        buildTransition:
+            vRouteElement.buildTransition ?? widget.buildTransition,
+        transitionDuration:
+            vRouteElement.transitionDuration ?? widget.transitionDuration,
+        reverseTransitionDuration: vRouteElement.reverseTransitionDuration ??
+            widget.reverseTransitionDuration,
         child: RouteElementWidget(
           stateKey: vRouteElement.stateKey,
-          child: vRouteElement.widget ?? Builder(builder: (context) => vRouteElement.widgetBuilder(context)),
+          child: vRouteElement.widget ??
+              Builder(
+                  builder: (context) => vRouteElement.widgetBuilder(context)),
           depth: index,
           pathParameters: localParameters,
           name: vRouteElement.name,
@@ -1101,11 +1126,12 @@ class VRouterState extends State<VRouter> {
 
     // Extract the path parameters from the url
     final match = newVRoutePathOfPath?.pathRegExp?.matchAsPrefix(newPath);
-    var newPathParameters =
-        (match != null) ? extract(newVRoutePathOfPath.parameters, match) : <String, String>{};
+    var newPathParameters = (match != null)
+        ? extract(newVRoutePathOfPath.parameters, match)
+        : <String, String>{};
     // Decode path parameters
-    newPathParameters =
-        newPathParameters.map((key, value) => MapEntry(key, Uri.decodeComponent(value)));
+    newPathParameters = newPathParameters
+        .map((key, value) => MapEntry(key, Uri.decodeComponent(value)));
 
     var shouldSaveHistoryState = false;
     var historyStatesToSave = {
@@ -1114,7 +1140,8 @@ class VRouterState extends State<VRouter> {
       '-1': vRoute?.key?.currentState?.historyState,
       for (var pages in flattenPages)
         '${pages.child.depth}':
-            pages.child.stateKey?.currentState?.historyState ?? pages.child.initialHistorySate,
+            pages.child.stateKey?.currentState?.historyState ??
+                pages.child.initialHistorySate,
     };
     String objectToSave;
     void saveHistoryState(String historyState) {
@@ -1135,7 +1162,7 @@ class VRouterState extends State<VRouter> {
         historyState: vRoute?.key?.currentState?.historyState,
         replaceHistoryState: (String _) =>
             throw 'replaceHistoryState cannot be called using this object.\n'
-                'If you want to save the state of the current route: Use saveHistoryState.\n'
+                'If you want to save the state of the current route: Use saveHistoryState in beforeLeave.\n'
                 'If you want to change the history state of the new route: call VRouterData.replaceHistoryState in afterEnter or afterUpdate.',
         pathParameters: vRoute?.pathParameters ?? {},
         queryParameters: vRoute?.queryParameters ?? {},
@@ -1145,7 +1172,7 @@ class VRouterState extends State<VRouter> {
         historyState: newState['-2'],
         replaceHistoryState: (String _) =>
             throw 'replaceHistoryState cannot be called using this object.\n'
-                'If you want to save the state of the current route: Use saveHistoryState.\n'
+                'If you want to save the state of the current route: Use saveHistoryState in beforeLeave.\n'
                 'If you want to change the history state of the new route: call VRouterData.replaceHistoryState in afterEnter or afterUpdate.',
         pathParameters: newPathParameters,
         queryParameters: queryParameters,
@@ -1155,8 +1182,9 @@ class VRouterState extends State<VRouter> {
     if (_url != null) {
       ///   1. beforeLeave in all deactivated VNavigationGuard
       for (var deactivatedPage in deactivatedPages) {
-        final vNavigationGuardMessages =
-            deactivatedPage.child.stateKey?.currentState?.vNavigationGuardMessages ?? [];
+        final vNavigationGuardMessages = deactivatedPage
+                .child.stateKey?.currentState?.vNavigationGuardMessages ??
+            [];
         for (var vNavigationGuardMessage in vNavigationGuardMessages) {
           if (vNavigationGuardMessage.vNavigationGuard.beforeLeave != null) {
             await vNavigationGuardMessage.vNavigationGuard
@@ -1169,7 +1197,8 @@ class VRouterState extends State<VRouter> {
         if (!vRedirector._shouldUpdate) {
           break;
         } else if (objectToSave != null &&
-            historyStatesToSave['${deactivatedPage.child.depth}'] != objectToSave) {
+            historyStatesToSave['${deactivatedPage.child.depth}'] !=
+                objectToSave) {
           historyStatesToSave['${deactivatedPage.child.depth}'] = objectToSave;
           objectToSave = null;
           shouldSaveHistoryState = true;
@@ -1185,17 +1214,18 @@ class VRouterState extends State<VRouter> {
             (BrowserHelpers.getHistorySerialCount() ?? 0) != serialCount) {
           ignoreNextBrowserCalls = true;
           BrowserHelpers.browserGo(serialCount - newSerialCount);
-          await BrowserHelpers.onBrowserPopState
-              .firstWhere((element) => BrowserHelpers.getHistorySerialCount() == serialCount);
+          await BrowserHelpers.onBrowserPopState.firstWhere((element) =>
+              BrowserHelpers.getHistorySerialCount() == serialCount);
           ignoreNextBrowserCalls = false;
         }
-        if (vRedirector._redirectFunction != null) vRedirector._redirectFunction();
+        if (vRedirector._redirectFunction != null)
+          vRedirector._redirectFunction();
         return;
       }
 
       ///   2. beforeLeave in the nest-most [VRouteElement] of the current route
       ///   saving the [VRoute] history state if needed
-      // Get the actual route
+      // Get the current route
       final vRoutePathOfPath = pathToRoutes.firstWhere(
           (_VRoutePath vRoutePathRegexp) =>
               vRoutePathRegexp.pathRegExp?.hasMatch(path) ?? false,
@@ -1220,11 +1250,12 @@ class VRouterState extends State<VRouter> {
               (BrowserHelpers.getHistorySerialCount() ?? 0) != serialCount) {
             ignoreNextBrowserCalls = true;
             BrowserHelpers.browserGo(serialCount - newSerialCount);
-            await BrowserHelpers.onBrowserPopState.firstWhere(
-                (element) => BrowserHelpers.getHistorySerialCount() == serialCount);
+            await BrowserHelpers.onBrowserPopState.firstWhere((element) =>
+                BrowserHelpers.getHistorySerialCount() == serialCount);
             ignoreNextBrowserCalls = false;
           }
-          if (vRedirector._redirectFunction != null) vRedirector._redirectFunction();
+          if (vRedirector._redirectFunction != null)
+            vRedirector._redirectFunction();
           return;
         }
       }
@@ -1247,11 +1278,12 @@ class VRouterState extends State<VRouter> {
               (BrowserHelpers.getHistorySerialCount() ?? 0) != serialCount) {
             ignoreNextBrowserCalls = true;
             BrowserHelpers.browserGo(serialCount - newSerialCount);
-            await BrowserHelpers.onBrowserPopState.firstWhere(
-                (element) => BrowserHelpers.getHistorySerialCount() == serialCount);
+            await BrowserHelpers.onBrowserPopState.firstWhere((element) =>
+                BrowserHelpers.getHistorySerialCount() == serialCount);
             ignoreNextBrowserCalls = false;
           }
-          if (vRedirector._redirectFunction != null) vRedirector._redirectFunction();
+          if (vRedirector._redirectFunction != null)
+            vRedirector._redirectFunction();
           return;
         }
       }
@@ -1271,11 +1303,12 @@ class VRouterState extends State<VRouter> {
               (BrowserHelpers.getHistorySerialCount() ?? 0) != serialCount) {
             ignoreNextBrowserCalls = true;
             BrowserHelpers.browserGo(serialCount - newSerialCount);
-            await BrowserHelpers.onBrowserPopState.firstWhere(
-                (element) => BrowserHelpers.getHistorySerialCount() == serialCount);
+            await BrowserHelpers.onBrowserPopState.firstWhere((element) =>
+                BrowserHelpers.getHistorySerialCount() == serialCount);
             ignoreNextBrowserCalls = false;
           }
-          if (vRedirector._redirectFunction != null) vRedirector._redirectFunction();
+          if (vRedirector._redirectFunction != null)
+            vRedirector._redirectFunction();
           return;
         }
       }
@@ -1297,18 +1330,21 @@ class VRouterState extends State<VRouter> {
               (BrowserHelpers.getHistorySerialCount() ?? 0) != serialCount) {
             ignoreNextBrowserCalls = true;
             BrowserHelpers.browserGo(serialCount - newSerialCount);
-            await BrowserHelpers.onBrowserPopState.firstWhere(
-                (element) => BrowserHelpers.getHistorySerialCount() == serialCount);
+            await BrowserHelpers.onBrowserPopState.firstWhere((element) =>
+                BrowserHelpers.getHistorySerialCount() == serialCount);
             ignoreNextBrowserCalls = false;
           }
-          if (vRedirector._redirectFunction != null) vRedirector._redirectFunction();
+          if (vRedirector._redirectFunction != null)
+            vRedirector._redirectFunction();
           return;
         }
       }
     }
 
     final oldSerialCount = serialCount;
-    if (shouldSaveHistoryState && path != null && historyStatesToSave.isNotEmpty) {
+    if (shouldSaveHistoryState &&
+        path != null &&
+        historyStatesToSave.isNotEmpty) {
       assert(
         kIsWeb,
         'Tried to store the state $historyStatesToSave while not on the web. State saving/restoration only work on the web.\n'
@@ -1324,14 +1360,14 @@ class VRouterState extends State<VRouter> {
       if (kIsWeb && fromBrowser && oldSerialCount != newSerialCount) {
         ignoreNextBrowserCalls = true;
         BrowserHelpers.browserGo(oldSerialCount - newSerialCount);
-        await BrowserHelpers.onBrowserPopState
-            .firstWhere((element) => BrowserHelpers.getHistorySerialCount() == oldSerialCount);
+        await BrowserHelpers.onBrowserPopState.firstWhere((element) =>
+            BrowserHelpers.getHistorySerialCount() == oldSerialCount);
       }
       BrowserHelpers.replaceHistoryState(jsonEncode(historyStatesToSave));
       if (kIsWeb && fromBrowser && oldSerialCount != newSerialCount) {
         BrowserHelpers.browserGo(newSerialCount - oldSerialCount);
-        await BrowserHelpers.onBrowserPopState
-            .firstWhere((element) => BrowserHelpers.getHistorySerialCount() == newSerialCount);
+        await BrowserHelpers.onBrowserPopState.firstWhere((element) =>
+            BrowserHelpers.getHistorySerialCount() == newSerialCount);
         ignoreNextBrowserCalls = false;
       }
     }
@@ -1364,9 +1400,11 @@ class VRouterState extends State<VRouter> {
         ignoreNextBrowserCalls = true;
         if (BrowserHelpers.getPathAndQuery(routerMode: widget.mode) != newUrl) {
           BrowserHelpers.pushReplacement(newUrl, routerMode: widget.mode);
-          if (BrowserHelpers.getPathAndQuery(routerMode: widget.mode) != newUrl) {
+          if (BrowserHelpers.getPathAndQuery(routerMode: widget.mode) !=
+              newUrl) {
             await BrowserHelpers.onBrowserPopState.firstWhere((element) =>
-                BrowserHelpers.getPathAndQuery(routerMode: widget.mode) == newUrl);
+                BrowserHelpers.getPathAndQuery(routerMode: widget.mode) ==
+                newUrl);
           }
         }
         BrowserHelpers.replaceHistoryState(jsonEncode(newState));
@@ -1411,11 +1449,17 @@ class VRouterState extends State<VRouter> {
     });
   }
 
-  /// See [VRouterData.onPop]
-  Future<void> _pop() async {
+  /// See [VRouterData.pop]
+  Future<void> _pop({
+    VRedirector vRedirector,
+    Map<String, String> queryParameters = const {},
+    String newState,
+  }) async {
     assert(_url != null);
 
-    final path = Uri.parse(_url).path;
+    // Instantiate VRedirector if null
+    // It might be not null if called from systemPop
+    vRedirector ??= _defaultPop(queryParameters: queryParameters);
 
     /// Call onPop in all active VNavigationGuards
     for (var page in flattenPages.reversed) {
@@ -1423,9 +1467,8 @@ class VRouterState extends State<VRouter> {
           page.child.stateKey?.currentState?.vNavigationGuardMessages ?? [];
       for (var vNavigationMessage in vNavigationMessages.reversed) {
         if (vNavigationMessage.vNavigationGuard.onPop != null) {
-          final shouldPop =
-              await vNavigationMessage.vNavigationGuard.onPop(vNavigationMessage.localContext);
-          if (!shouldPop) {
+          await vNavigationMessage.vNavigationGuard.onPop(vRedirector);
+          if (!vRedirector.shouldUpdate) {
             return;
           }
         }
@@ -1433,6 +1476,9 @@ class VRouterState extends State<VRouter> {
     }
 
     /// Call onPop of the nested-most VRouteElement
+    // Get the current path
+    final path = Uri.parse(_url).path;
+
     // Get the current route
     final vRoutePathLocals = pathToRoutes
         .firstWhere(
@@ -1445,29 +1491,41 @@ class VRouterState extends State<VRouter> {
     // having onPopPage implemented
     final vRouteElement = vRoutePathLocals.last;
     if (vRouteElement.onPop != null) {
-      final shouldPop = await vRouteElement.onPop(_vRouterInformationContext);
-      if (!shouldPop) {
+      await vRouteElement.onPop(vRedirector);
+      if (!vRedirector.shouldUpdate) {
         return;
       }
     }
 
     /// Call onPop of VRouter
     if (widget.onPop != null) {
-      final shouldPop = await widget.onPop(_vRouterInformationContext);
-      if (!shouldPop) {
+      await widget.onPop(vRedirector);
+      if (!vRedirector.shouldUpdate) {
         return;
       }
     }
 
-    /// Call default onPop
-    _defaultPop();
+    /// Update the url to the one found in [_defaultPop]
+    if (vRedirector != null) {
+      _updateUrl(vRedirector.to,
+          queryParameters: queryParameters, newState: {'-2': newState});
+    } else if (!kIsWeb) {
+      // If we didn't find a url to go to, we are at the start of the stack
+      // so we close the app on mobile
+      MoveToBackground.moveTaskToBack();
+    }
   }
 
   /// See [VRouterData.systemPop]
-  Future<void> _systemPop() async {
+  Future<void> _systemPop({
+    Map<String, String> queryParameters = const {},
+    String newState,
+  }) async {
     assert(_url != null);
 
-    final path = Uri.parse(_url).path;
+    // Instantiate VRedirector if null
+    // It might be not null if called from systemPop
+    final vRedirector = _defaultPop(queryParameters: queryParameters);
 
     /// Call onSystemPop in all active VNavigationGuards
     for (var page in flattenPages.reversed) {
@@ -1475,9 +1533,8 @@ class VRouterState extends State<VRouter> {
           page.child.stateKey?.currentState?.vNavigationGuardMessages ?? [];
       for (var vNavigationMessage in vNavigationMessages.reversed) {
         if (vNavigationMessage.vNavigationGuard.onSystemPop != null) {
-          final shouldPop = await vNavigationMessage.vNavigationGuard
-              .onSystemPop(vNavigationMessage.localContext);
-          if (!shouldPop) {
+          await vNavigationMessage.vNavigationGuard.onSystemPop(vRedirector);
+          if (!vRedirector.shouldUpdate) {
             return;
           }
         }
@@ -1485,6 +1542,9 @@ class VRouterState extends State<VRouter> {
     }
 
     /// Call onSystemPop of the nested-most VRouteElement
+    // Get the current path
+    final path = Uri.parse(_url).path;
+
     // Get the current route
     final vRoutePathLocals = pathToRoutes
         .firstWhere(
@@ -1498,8 +1558,8 @@ class VRouterState extends State<VRouter> {
     final vRouteElement = vRoutePathLocals.last;
     if (vRouteElement.onSystemPop != null) {
       // If we did find a VRouteClass, call onSystemPopPage
-      final shouldPop = await vRouteElement.onSystemPop(_vRouterInformationContext);
-      if (!shouldPop) {
+      await vRouteElement.onSystemPop(vRedirector);
+      if (!vRedirector.shouldUpdate) {
         return;
       }
     }
@@ -1507,8 +1567,8 @@ class VRouterState extends State<VRouter> {
     /// Call VRouter onSystemPop
     if (widget.onSystemPop != null) {
       // Call VRouter.onSystemPopPage if implemented
-      final shouldPop = await widget.onSystemPop(_vRouterInformationContext);
-      if (!shouldPop) {
+      await widget.onSystemPop(vRedirector);
+      if (!vRedirector.shouldUpdate) {
         return;
       }
     }
@@ -1517,13 +1577,10 @@ class VRouterState extends State<VRouter> {
     await _pop();
   }
 
-  /// This is the default behaviour of pop which is call if no one
-  /// handled a pop event or a system pop event.
-  /// This methods find the closest [VRouteElement] where [VRouteElement.isChild]
-  /// is false, then finds the url corresponding to this subroute and
-  /// call [_updateUrl] with this url.
-  /// If no such [VRouteElement] is found, we do nothing on the web but
-  /// put the application on the background on mobile.
+  /// This finds new url when a pop event occurs by finding the closest [VRouteElement]
+  /// where [VRouteElement.isChild] is false and [VRouteElement.path] is not null
+  /// It returns a [VRedirector] with the newVRouteData corresponding to the found path.
+  /// If no such [VRouteElement] is found, newVRouteData is null
   ///
   /// We also try to preserve path parameters if possible
   /// For example
@@ -1541,7 +1598,10 @@ class VRouterState extends State<VRouter> {
   ///                               when the call comes from the system
   ///   * [VRouter.onSystemPop] to override this behaviour on a global level
   ///                               when the call comes from the system
-  void _defaultPop() {
+  VRedirector _defaultPop({
+    Map<String, String> queryParameters = const {},
+    String newState,
+  }) {
     assert(_url != null);
     final path = Uri.parse(_url).path;
 
@@ -1559,28 +1619,29 @@ class VRouterState extends State<VRouter> {
     vRouteElements.removeLast();
 
     // Find the vRouteElements where vRoute.isChild of the last element is false
-    while (vRouteElements.isNotEmpty && vRouteElements.last.isChild) {
+    while (vRouteElements.isNotEmpty &&
+        (vRouteElements.last.isChild || vRouteElements.last.path == null)) {
       vRouteElements.removeLast();
     }
 
+    // This VRouteData will be not null if we find a route to go to
+    VRouteData newVRouteData;
+
+    // This url will be not null if we find a route to go to
+    String newUrl;
+
     if (vRouteElements.isNotEmpty) {
-      // Get the new path
-      final newRawPath = vRouteElements.fold<String>(
-        '/',
-        (previousValue, VRouteElement vRouteElement) {
-          if (vRouteElement.path == null) {
-            return previousValue;
-          } else if (vRouteElement.path.startsWith('/')) {
-            return vRouteElement.path;
-          } else {
-            return previousValue + '/' + vRouteElement.path;
-          }
-        },
+      // Get the VRoutePath from the vRouteElements
+      final newVRoutePath = pathToRoutes.firstWhere(
+        (vRoutePath) => listEquals(vRoutePath.vRouteElements, vRouteElements),
       );
-      final newPathRegExp = pathToRegExp(newRawPath, prefix: true);
+
+      // Get the new pathRegexp as a prefix regExp
+      final newPathRegExpPrefix =
+          pathToRegExp(newVRoutePath.path, prefix: true);
 
       // Extract the newRawPath from the path using the newRawPath
-      final match = newPathRegExp.matchAsPrefix(path);
+      final match = newPathRegExpPrefix.matchAsPrefix(path);
       String newPath;
       if (match != null) {
         newPath = path.substring(match.start, match.end);
@@ -1590,34 +1651,60 @@ class VRouterState extends State<VRouter> {
         // This means that any parameter won't be able to be restored,
         // this is expected since we have no way to deduce those parameters
         // from the current path
-        newPath = newRawPath;
+        newPath = newVRoutePath.path;
       }
 
-      // We keep the queryParameters
-      final newUrl = Uri(
-              path: newPath,
-              queryParameters: (vRoute?.queryParameters?.isNotEmpty ?? false)
-                  ? vRoute.queryParameters
-                  : null)
-          .toString();
+      // Integrate the given query parameters
+      newUrl = Uri(
+        path: newPath,
+        queryParameters: (queryParameters.isNotEmpty) ? queryParameters : null,
+      ).toString();
 
-      // Update the url
-      _updateUrl(newUrl);
-      return;
+      // Extract the path parameters from the url
+      final newMatch = newVRoutePath.pathRegExp.matchAsPrefix(newPath);
+      var newPathParameters = (newMatch != null)
+          ? extract(newVRoutePath.parameters, newMatch)
+          : <String, String>{};
+      // Decode path parameters
+      newPathParameters = newPathParameters
+          .map((key, value) => MapEntry(key, Uri.decodeComponent(value)));
+
+      newVRouteData = VRouteData(
+        child: Container(),
+        historyState: newState,
+        replaceHistoryState: (String _) =>
+            throw 'replaceHistoryState cannot be called using this object.\n'
+                'If you want to save the state of the current route: Use saveHistoryState in beforeLeave.\n'
+                'If you want to change the history state of the new route: call VRouterData.replaceHistoryState in afterEnter or afterUpdate.',
+        pathParameters: newPathParameters,
+        queryParameters: queryParameters,
+      );
     }
 
-    // If we didn't find any route, we are at the bottom of the stack
-    // So if we are on mobile we move the app to the background
-    if (!kIsWeb) {
-      MoveToBackground.moveTaskToBack();
-    }
+    return VRedirector(
+      context: _vRouterInformationContext,
+      from: _url,
+      to: newUrl,
+      previousVRouteData: VRouteData(
+        child: Container(),
+        historyState: vRoute?.key?.currentState?.historyState,
+        replaceHistoryState: (String _) =>
+            throw 'replaceHistoryState cannot be called using this object.\n'
+                'If you want to save the state of the current route: Use saveHistoryState in beforeLeave.\n'
+                'If you want to change the history state of the new route: call VRouterData.replaceHistoryState in afterEnter or afterUpdate.',
+        pathParameters: vRoute?.pathParameters ?? {},
+        queryParameters: vRoute?.queryParameters ?? {},
+      ),
+      newVRouteData: newVRouteData,
+    );
   }
 
   /// See [VRouterData.replaceHistoryState]
   void _replaceHistoryState(String newRouterState) {
     if (kIsWeb) {
       final historyState = BrowserHelpers.getHistoryState() ?? '{}';
-      final historyStateMap = Map<String, String>.from(jsonDecode(historyState));
+      final historyStateMap =
+          Map<String, String>.from(jsonDecode(historyState));
       historyStateMap['-2'] = newRouterState;
       final newHistoryState = jsonEncode(historyStateMap);
       BrowserHelpers.replaceHistoryState(newHistoryState);
@@ -1649,13 +1736,16 @@ class VRouterState extends State<VRouter> {
 
     ///   1. beforeLeave in all deactivated vRouteElement
     for (var deactivatedPage in flattenPages) {
-      final vNavigationMessages =
-          deactivatedPage.child.stateKey?.currentState?.vNavigationGuardMessages ?? [];
+      final vNavigationMessages = deactivatedPage
+              .child.stateKey?.currentState?.vNavigationGuardMessages ??
+          [];
       for (var vNavigationMessage in vNavigationMessages) {
         if (vNavigationMessage.vNavigationGuard.beforeLeave != null) {
-          await vNavigationMessage.vNavigationGuard.beforeLeave(null, saveHistoryState);
+          await vNavigationMessage.vNavigationGuard
+              .beforeLeave(null, saveHistoryState);
           if (objectToSave != null) {
-            historyStatesToSave['${deactivatedPage.child.depth}'] = objectToSave;
+            historyStatesToSave['${deactivatedPage.child.depth}'] =
+                objectToSave;
             objectToSave = null;
             shouldSaveHistoryState = true;
           }
@@ -1666,7 +1756,8 @@ class VRouterState extends State<VRouter> {
     ///   2. beforeLeave in the nest-most [VRouteElement] of the current route
     // Get the actual route
     final vRoutePathOfPath = pathToRoutes.firstWhere(
-        (_VRoutePath vRoutePathRegexp) => vRoutePathRegexp.pathRegExp?.hasMatch(path) ?? false,
+        (_VRoutePath vRoutePathRegexp) =>
+            vRoutePathRegexp.pathRegExp?.hasMatch(path) ?? false,
         orElse: () => throw InvalidUrlException(url: path));
 
     // Call the nest-most VRouteClass of the current route
@@ -1715,8 +1806,14 @@ class VRouterData extends InheritedWidget {
     Map<String, String> newState,
     bool isReplacement,
   }) _updateUrlFromName;
-  final Future<void> Function() _pop;
-  final Future<void> Function() _systemPop;
+  final Future<void> Function({
+    Map<String, String> queryParameters,
+    String newState,
+  }) _pop;
+  final Future<void> Function({
+    Map<String, String> queryParameters,
+    String newState,
+  }) _systemPop;
   final void Function(String historyState) _replaceHistoryState;
 
   VRouterData({
@@ -1746,9 +1843,17 @@ class VRouterData extends InheritedWidget {
     })
             updateUrlFromName,
     @required
-        Future<void> Function() pop,
+        Future<void> Function({
+      Map<String, String> queryParameters,
+      String newState,
+    })
+            pop,
     @required
-        Future<void> Function() systemPop,
+        Future<void> Function({
+      Map<String, String> queryParameters,
+      String newState,
+    })
+            systemPop,
     @required
         void Function(String historyState) replaceHistoryState,
   })  : _updateUrl = updateUrl,
@@ -1816,7 +1921,8 @@ class VRouterData extends InheritedWidget {
       newUrl = currentPath + '/$newUrl';
     }
 
-    _updateUrl(newUrl, queryParameters: queryParameters, newState: {'-2': routerState});
+    _updateUrl(newUrl,
+        queryParameters: queryParameters, newState: {'-2': routerState});
   }
 
   /// Updates the url given a [VRouteElement] name
@@ -1868,7 +1974,8 @@ class VRouterData extends InheritedWidget {
   }) {
     // If not on the web, this is the same as push
     if (!kIsWeb) {
-      return push(newUrl, queryParameters: queryParameters, routerState: routerState);
+      return push(newUrl,
+          queryParameters: queryParameters, routerState: routerState);
     }
 
     if (!newUrl.startsWith('/')) {
@@ -1933,8 +2040,13 @@ class VRouterData extends InheritedWidget {
   ///   3. onPop is called in [VRouter]
   ///   4. Default behaviour of pop is called: [VRouterState._defaultPop]
   ///
-  /// If any of the step return false, then the pop cycle stops
-  void pop(BuildContext context) => _pop();
+  /// In any of the above steps, we can use [vRedirector] if you want to redirect or
+  /// stop the navigation
+  void pop({
+    Map<String, String> queryParameters = const {},
+    String routerState,
+  }) =>
+      _pop(queryParameters: queryParameters, newState: routerState);
 
   /// Starts a systemPop cycle
   ///
@@ -1944,14 +2056,21 @@ class VRouterData extends InheritedWidget {
   ///   3. onSystemPop is called in VRouter
   ///   4. [pop] is called
   ///
-  /// If any of the step return false, then the pop cycle stops
-  Future<void> systemPop(BuildContext context) => _systemPop();
+  /// In any of the above steps, we can use [vRedirector] if you want to redirect or
+  /// stop the navigation
+  Future<void> systemPop({
+    Map<String, String> queryParameters = const {},
+    String routerState,
+  }) =>
+      _systemPop(queryParameters: queryParameters, newState: routerState);
 
   /// This replaces the current history state of [VRouterData] with given one
-  void replaceHistoryState(String historyState) => _replaceHistoryState(historyState);
+  void replaceHistoryState(String historyState) =>
+      _replaceHistoryState(historyState);
 
   static VRouterData of(BuildContext context) {
-    final vRouterData = context.dependOnInheritedWidgetOfExactType<VRouterData>();
+    final vRouterData =
+        context.dependOnInheritedWidgetOfExactType<VRouterData>();
     if (vRouterData == null) {
       throw FlutterError(
           'VRouterData.of(context) was called with a context which does not contain a VRouter.\n'