+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## [0.0.11] - 20/02/2021
+
+* give a vRedirector when handling pop events
+
 ## [0.0.10] - 19/02/2021
 
 * add widgetBuilder for VStacked and VChild
 
 ## [0.0.9] - 18/02/2021
 
-* add vRouteData in beforeLeave and beforeEnter
+* add vRedirector in beforeLeave and beforeEnter
 * fix error when replacing path parameters in pushNamed
 * fix error with pushReplace on the web